refactor: EventCard
